+++ test/runtests.jl
@@ -1,2 +1,2 @@
-module lowrankmatrix_tests include("lowrankmatrix.jl") end
+module factoredmatrix_tests include("factoredmatrix.jl") end
 module aqua_tests include("aqua.jl") end